Modifica componentes para reutilizacion
